frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.0.1"
description = "Web Frontend for HEMS"
authors = ["Hydro-Québec <hydroquebec@hydroquebec.com>"]
readme = "README.md"
packages = [{include = "frontend", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
pyyaml = "^6.0.2"
requests = "^2.32.3"
redis = "^5.2.1"
jinja2 = "^3.1.2"
common = { path = "../common", develop = true }

[tool.poetry.group.dev.dependencies]
types-requests = "^2.32.0.20250328"
types-pyyaml = "^6.0.12.20250326"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
